+++ tests/test_hamiltonian.py
@@ -1,0 +1,8 @@
+"""Tests the evaluation of the eigenvalues and eigenvectors of 1D-quantum potentials."""
+
+import pytest
+from basis.hamiltonian import Hamiltonian
+import numpy as np
+import sys
+
+